Staging

+++ src/components/AttachmentThumbnail/index.tsx
@@ -11,6 +11,10 @@ import { useData } from 'src/StateProvider/Provider';
 import PreviewIcon from '@material-ui/icons/Visibility';
 import _ from 'lodash';
 import ConfirmationDialog from 'src/components/Helpers/ConfirmationDialog';
+import moment from 'moment';
+import { dateTimeFormat } from 'src/constants/helpers';
+import axios from 'axios';
+import mimeDb from 'mime-db';
 
 const fileIcons = [
   {
@@ -44,7 +48,6 @@ const fileIcons = [
 ];
 
 const AttachmentThumbnail = ({ attachments, handleDeleteAttachment, canEdit }) => {
-  console.log(attachments);
   const {
     state: { permissions }
   }: any = useData();
@@ -57,12 +60,18 @@ const AttachmentThumbnail = ({ attachments, handleDeleteAttachment, canEdit }) =
 
   // GET ATTACHMENT ICON
   const getFileIconSrc = (file) => {
-    if (file) {
+    if (file.contentType) {
+      let extension = `.${mimeDb[file.contentType].extensions[0]}`;
+      let data = fileIcons.find((o) => o.extensions.indexOf(extension) >= 0);
+      if (data && data?.source) return data.source;
+    } else if (file) {
       let extension = file.substring(file.lastIndexOf('.')).toLowerCase();
       let data = fileIcons.find((o) => o.extensions.indexOf(extension) >= 0);
       if (data && data?.source) return data.source;
     }
+    return '';
   };
+
   // VIEW ATTACHMENT
   const viewPdf = (event, file) => {
     if (event) {
@@ -116,37 +125,70 @@ const AttachmentThumbnail = ({ attachments, handleDeleteAttachment, canEdit }) =
         message: `Downloading, Please wait...`
       });
     }
-    setDownloadProgress(0);
     setIsDownloading(true);
-    axiosInstance()
-      .get(`user/download?fileName=${file.url}`, {
-        responseType: 'blob',
-        onDownloadProgress: (progressEvent) => {
-          let percentCompleted = Math.floor((progressEvent.loaded * 100) / progressEvent.total);
-          setDownloadProgress(percentCompleted);
+    setDownloadProgress(0);
 
-          if (percentCompleted === 100) {
-            toastConfig.setToastConfig({ open: true, type: 'success', message: 'File downloaded successfully.' });
-            setTimeout(() => {
-              setDownloadProgress(0);
-              setIsDownloading(false);
-            }, 2000);
+    if (file.url) {
+      axiosInstance()
+        .get(`user/download?fileName=${file.url}`, {
+          responseType: 'blob',
+          onDownloadProgress: (progressEvent) => {
+            let percentCompleted = Math.floor((progressEvent.loaded * 100) / progressEvent.total);
+            setDownloadProgress(percentCompleted);
+
+            if (percentCompleted === 100) {
+              toastConfig.setToastConfig({ open: true, type: 'success', message: 'File downloaded successfully.' });
+              setTimeout(() => {
+                setDownloadProgress(0);
+                setIsDownloading(false);
+              }, 2000);
+            }
           }
-        }
-      })
-      .then(({ data }) => {
-        const url = window.URL.createObjectURL(new Blob([data]));
-        const link = document.createElement('a');
-        link.href = url;
-        link.setAttribute('download', file.url);
-        document.body.appendChild(link);
-        link.click();
-        setTimeout(() => setIsDownloading(false), 2000);
-      })
-      .catch((err) => {
-        toastConfig.setToastConfig(err);
-        setIsDownloading(false);
-      });
+        })
+        .then(({ data }) => {
+          const url = window.URL.createObjectURL(new Blob([data]));
+          const link = document.createElement('a');
+          link.href = url;
+          link.setAttribute('download', file.url);
+          document.body.appendChild(link);
+          link.click();
+          setTimeout(() => setIsDownloading(false), 2000);
+        })
+        .catch((err) => {
+          toastConfig.setToastConfig(err);
+          setIsDownloading(false);
+        });
+    } else {
+      axios
+        .get(file, {
+          responseType: 'blob',
+          onDownloadProgress: (progressEvent) => {
+            let percentCompleted = Math.floor((progressEvent.loaded * 100) / progressEvent.total);
+            setDownloadProgress(percentCompleted);
+
+            if (percentCompleted === 100) {
+              toastConfig.setToastConfig({ open: true, type: 'success', message: 'File downloaded successfully.' });
+              setTimeout(() => {
+                setDownloadProgress(0);
+                setIsDownloading(false);
+              }, 2000);
+            }
+          }
+        })
+        .then((data) => {
+          const url = window.URL.createObjectURL(new Blob([data.data]));
+          const link = document.createElement('a');
+          link.href = url;
+          link.setAttribute('download', file?.substring(file.lastIndexOf('/') + 1));
+          document.body.appendChild(link);
+          link.click();
+          setTimeout(() => setIsDownloading(false), 2000);
+        })
+        .catch((err) => {
+          toastConfig.setToastConfig(err);
+          setIsDownloading(false);
+        });
+    }
   };
 
   return (
@@ -156,76 +198,85 @@ const AttachmentThumbnail = ({ attachments, handleDeleteAttachment, canEdit }) =
           <>
             {attachments.map((attachment, i) => {
               return (
-                <>
-                  <Grid item key={i} sm={3} xs={3} md={3} xl={3}>
-                    <Paper className={emailStyles.fileContainer}>
-                      <img src={getFileIconSrc(attachment.url)} className={emailStyles.file} alt="attchment" />
-                      <Typography noWrap variant="body2">
+                <Grid item key={i} sm={3} xs={3} md={3} xl={3} style={{ maxWidth: '150px' }}>
+                  <Paper className={emailStyles.fileContainer}>
+                    <img
+                      src={getFileIconSrc(attachment?.contentType ? attachment?.contentType : attachment.url ? attachment.url : attachment)}
+                      className={emailStyles.file}
+                      alt="attchment"
+                    />
+                    <Typography noWrap variant="body2">
+                      {attachment
+                        ? attachment?.name
+                          ? attachment?.name
+                          : attachment?.url?.substring(attachment.url.lastIndexOf('/') + 1)
+                          ? attachment?.url?.substring(attachment.url.lastIndexOf('/') + 1)
+                          : attachment.substring(attachment.lastIndexOf('/') + 1)
+                        : 'attachment'}
+                    </Typography>
+                    {attachment?.date && <Typography variant="body2">{moment(attachment?.date)?.format(dateTimeFormat)}</Typography>}
+                    <div className={emailStyles.fileOverlay}>
+                      <Typography
+                        variant="subtitle2"
+                        title={
+                          attachment
+                            ? attachment?.name
+                              ? attachment?.name
+                              : attachment?.url?.substring(attachment.url.lastIndexOf('/') + 1)
+                              ? attachment?.url?.substring(attachment.url.lastIndexOf('/') + 1)
+                              : attachment?.substring(attachment.lastIndexOf('/') + 1)
+                            : 'attachment'
+                        }
+                      >
                         {attachment
                           ? attachment?.name
                             ? attachment?.name
-                            : attachment.url.substring(attachment.url.lastIndexOf('/') + 1)
+                            : attachment?.url?.substring(attachment.url.lastIndexOf('/') + 1)
+                            ? attachment?.url?.substring(attachment.url.lastIndexOf('/') + 1)
+                            : attachment?.substring(attachment.lastIndexOf('/') + 1)
                           : 'attachment'}
                       </Typography>
-                      <div className={emailStyles.fileOverlay}>
-                        <Typography variant="subtitle2">
-                          {attachment
-                            ? attachment?.name
-                              ? attachment?.name
-                              : attachment.url.substring(attachment.url.lastIndexOf('/') + 1)
-                            : 'attachment'}
-                        </Typography>
-                        <div style={{ display: 'flex', justifyContent: 'space-between', width: '50%', float: 'right', bottom: '0' }}>
-                          <Tooltip title="Download">
-                            <IconButton onClick={(event) => downloadFile(event, attachment)} style={{ paddingBottom: '1px' }}>
-                              {<GetAppIcon />}
+                      <div style={{ display: 'flex', justifyContent: 'space-between' }}>
+                        <Tooltip title="Download" placement="top">
+                          <IconButton size={'small'} onClick={(event) => downloadFile(event, attachment)} style={{ paddingBottom: '1px' }}>
+                            {<GetAppIcon />}
+                          </IconButton>
+                        </Tooltip>
+                        {_.endsWith(attachment?.url, '.pdf') && (
+                          <Tooltip title="Preview" placement="top">
+                            <IconButton
+                              size={'small'}
+                              onClick={(e) => {
+                                viewPdf(e, attachment.url);
+                              }}
+                            >
+                              <PreviewIcon color="primary" />
                             </IconButton>
                           </Tooltip>
-                          {_.endsWith(attachment?.url, '.pdf') && (
-                            <Tooltip title="Preview">
-                              <IconButton>
-                                <PreviewIcon
-                                  color="primary"
-                                  onClick={(e) => {
-                                    viewPdf(e, attachment.url);
-                                  }}
-                                />
-                              </IconButton>
-                            </Tooltip>
-                          )}
-                          {canEdit && permissions.attachment.isDelete ? (
-                            <Tooltip title="Delete">
-                              <IconButton>
-                                {
-                                  <DeleteIcon
-                                    color="error"
-                                    onClick={() => {
-                                      setShowConfirmationDialog(true);
-                                      setAttachemnetToDelete(attachment);
-                                    }}
-                                  />
-                                }
-                              </IconButton>
-                            </Tooltip>
-                          ) : (
-                            <Tooltip
-                              className="cursor-stop"
-                              title={
-                                permissions.quoteBuilder.isDelete
-                                  ? 'Signed quote attachments can not be deleted'
-                                  : "You don't have permissions to delete attachment"
-                              }
+                        )}
+                        {canEdit && permissions.attachment.isDelete ? (
+                          <Tooltip title="Delete" placement="top">
+                            <IconButton
+                              size={'small'}
+                              onClick={() => {
+                                setShowConfirmationDialog(true);
+                                setAttachemnetToDelete(attachment);
+                              }}
                             >
-                              <IconButton>
-                                <DeleteIcon color="disabled" />
-                              </IconButton>
-                            </Tooltip>
-                          )}
-                        </div>
+                              {<DeleteIcon color="error" />}
+                            </IconButton>
+                          </Tooltip>
+                        ) : (
+                          <Tooltip className="cursor-stop" title={"You don't have permissions to delete attachment"}>
+                            <IconButton size={'small'}>
+                              <DeleteIcon color="disabled" />
+                            </IconButton>
+                          </Tooltip>
+                        )}
                       </div>
-                    </Paper>
-                  </Grid>
-                </>
+                    </div>
+                  </Paper>
+                </Grid>
               );
             })}
           </>
@@ -240,7 +291,9 @@ const AttachmentThumbnail = ({ attachments, handleDeleteAttachment, canEdit }) =
           }}
           onOk={() => {
             setShowConfirmationDialog(false);
-            handleDeleteAttachment(attachmentToDelete);
+            if (handleDeleteAttachment) {
+              handleDeleteAttachment(attachmentToDelete);
+            }
           }}
         />
       )}
